requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
sqlparse==0.5.3
pytest==8.3.3
pytest-django==4.9.0
